Accomodate changes for the protocol unmodified validator
